move exit function to builtins.c
